move resources to specific packages
