+++ resources/bluima_sentence_detector_genia/src/main/java/ch/epfl/bbp/nlp/res/sentence/GeniaResource.java
@@ -1,4 +1,6 @@
-package ch.epfl.bbp.nlp;
+package ch.epfl.bbp.nlp.res.sentence;
+
+import ch.epfl.bbp.nlp.ModelResource;
 
 /**
  * Genia model for sentence splitter annotator
