+++ tests/bluima_dummymodelproxy/src/test/java/ch/epfl/bbp/nlp/TestGeniaResource.java
@@ -3,11 +3,14 @@ package ch.epfl.bbp.nlp;
 import org.junit.Assert;
 import org.junit.Test;
 
+import ch.epfl.bbp.nlp.res.sentence.GeniaResource;
+
 public class TestGeniaResource {
 
     @Test
     public final void testGetModelFromName() throws ModelProxyException {
-        ModelStream stream = ModelProxy.getStream("ch.epfl.bbp.nlp.GeniaResource");
+        ModelStream stream = ModelProxy
+                .getStream("ch.epfl.bbp.nlp.res.sentence.GeniaResource");
         Assert.assertNotNull("Resource is not null", stream);
     }
