+++ tests/bluima_dummymodelproxy/src/test/java/ch/epfl/bbp/nlp/TestPennBioResource.java
@@ -3,11 +3,14 @@ package ch.epfl.bbp.nlp;
 import org.junit.Assert;
 import org.junit.Test;
 
+import ch.epfl.bbp.nlp.res.sentence.PennBioResource;
+
 public class TestPennBioResource {
 
     @Test
     public final void testGetModelFromName() throws ModelProxyException {
-        ModelStream stream = ModelProxy.getStream("ch.epfl.bbp.nlp.PennBioResource");
+        ModelStream stream = ModelProxy
+                .getStream("ch.epfl.bbp.nlp.res.sentence.PennBioResource");
         Assert.assertNotNull("Resource is not null", stream);
     }
